Reorganise internal file structure
