+++ lib/src/commands/group.dart
@@ -18,12 +18,12 @@ import 'package:logging/logging.dart';
 import 'package:meta/meta.dart';
 import 'package:nyxx_interactions/nyxx_interactions.dart';
 
-import 'checks.dart';
+import '../checks/checks.dart';
 import 'command.dart';
-import 'commands.dart';
-import 'context.dart';
-import 'errors.dart';
-import 'view.dart';
+import '../commands.dart';
+import '../context/context.dart';
+import '../errors.dart';
+import '../util/view.dart';
 
 /// A [Group] is a collection of commands. This mixin implements that functionality.
 ///
